Make the code easier to type check
Type checkers like `mypy` do not actually run the code they are checking.
This makes certain legal constructs problematic for them.

In this case, constructing the `namedtuple` from the keys of a
dictionary makes it impossible for the type checker to know the names of
the values of the tuple.

Instead, we use a `typing.NamedTuple` and list the fields manually. With
this change, the type checker knows the names of the fields and their
expected type.